Refactor catalog repository with robust error handling
